mlp/requirements.txt
torch>=1.7.0
torchvision>=0.8.1
ultralytics>=8.0.0
opencv-python==4.5.4.60
matplotlib==3.4.3
tensorboard==2.7.0
numpy==1.21.4
tqdm>=4.64.0  # Обновлено для совместимости с ultralytics
scikit-learn==0.24.2
pyyaml==5.4.1
Pillow==8.4.0
